Added model for Item
